+++ README.md
@@ -1,2 +1,2 @@
-# courtsdk
+# Courtsdk
 Ferramentas para desenvolver Crawlers/Scrapers de forma sequencial ou concorrente.
