+++ helpers_test.go
@@ -23,3 +23,21 @@ func TestGetEnvString(t *testing.T) {
 		t.Errorf("Env var was incorrect\n Got:  %s\n Want: %s", envValue, testValue)
 	}
 }
+
+func TestGenerateMD5(t *testing.T) {
+	expectedHash := "86fb269d190d2c85f6e0468ceca42a20"
+	phrase := "Hello world!"
+	generatedHash := GenerateMD5(&phrase)
+	if generatedHash != expectedHash {
+		t.Errorf("MD5 checksum different from expected\n Got:  %s\n Want: %s", generatedHash, expectedHash)
+	}
+}
+
+func TestGetElasticMapping(t *testing.T) {
+	expectedHash := "c58d3987aa778d466ac1c2fcff1ac945"
+	mapping := GetElasticMapping()
+	generatedHash := GenerateMD5(&mapping)
+	if generatedHash != expectedHash {
+		t.Errorf("Elasticsearch mapping was different from test model\n Got:  %s\n Want: %s", generatedHash, expectedHash)
+	}
+}
